Updated repositories

+++ src/main/java/ua/casten/bowling/model/repository/FrameRepository.java
@@ -7,7 +7,7 @@ import ua.casten.bowling.model.Frame;
 import java.util.List;
 
 @Repository
-public interface FrameRepository extends JpaRepository<Frame, Long> {
+public interface FrameRepository extends JpaRepository<Frame, Integer> {
 
     List<Frame> findFramesByGameId(int gameId);
